+++ docs/make.jl
@@ -14,7 +14,8 @@ PAGES = [
         "uncertain_values/uncertainvalues_fitted.md",
         "uncertain_values/uncertainvalues_theoreticaldistributions.md",
         "uncertain_values/certainvalue.md",
-        "uncertain_values/populations.md"
+        "uncertain_values/populations.md",
+        "uncertain_values/merging.md"
     ],
 	"Uncertain datasets" => [
         "uncertain_datasets/uncertain_datasets_overview.md",